replace new bestuursorgaan page with modal
