+++ app/router.js
@@ -35,7 +35,6 @@ Router.map(function () {
       this.route('mandaten');
     });
     this.route('beheer', function () {
-      this.route('new');
       this.route('edit', { path: '/:orgaan_id/edit' });
     });
     this.route('fracties');
