--- app/templates/organen/beheer/new.hbs
@@ -1,25 +1,0 @@
-{{page-title "Nieuw Bestuursorgaan"}}
-<AuToolbar @border="bottom" @size="large" as |Group|>
-  <Group>
-    <AuHeading @skin="2">Voeg bestuursorgaan toe</AuHeading>
-  </Group>
-</AuToolbar>
-
-<div class="au-c-body-container au-c-body-container--scroll">
-  <div class="au-o-box">
-    <Form::NewInstance
-      @onCreate={{this.onCreate}}
-      @form={{this.model.form}}
-      @buildSourceTtl={{this.buildSourceTtl}}
-    />
-  </div>
-</div>
-
-<AuToolbar @border="top" @size="large" as |Group|>
-  <Group>
-    <Form::CancelWithConfirm
-      @cancel={{this.cancel}}
-    >Annuleer</Form::CancelWithConfirm>
-  </Group>
-</AuToolbar>
-{{outlet}}
